langchain-poc/traceloop/requirements.txt
langchain
langchain-community
langchain-aws
opentelemetry-sdk
opentelemetry-api
opentelemetry-semantic-conventions
python-dotenv
traceloop-sdk
botocore
boto3
setuptools
aws-opentelemetry-distro-genai-beta==0.1.0b6
fastapi
uvicorn
